requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Data & Finance
yfinance>=0.2.33
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
httpx>=0.25.1
inflect>=7.0.0

# Utilities
cachetools>=5.3.1
loguru>=0.7.2
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
